Add ICMP and UDP support for NAT and loadbalancer
