+++ include/dp_nat.h
@@ -27,6 +27,7 @@ void dp_del_vm_dnat_ip(uint32_t d_ip, uint32_t vni);
 bool dp_is_ip_dnatted(uint32_t d_ip, uint32_t vni);
 uint32_t dp_get_vm_dnat_ip(uint32_t d_ip, uint32_t vni);
 void dp_set_vm_dnat_ip(uint32_t d_ip, uint32_t vm_ip, uint32_t vni);
+void dp_nat_chg_ip(struct dp_flow *df_ptr, struct rte_ipv4_hdr *ipv4_hdr);
 
 #ifdef __cplusplus
 }
