+++ src/nodes/snat_node.c
@@ -27,7 +27,6 @@ static int snat_node_init(const struct rte_graph *graph, struct rte_node *node)
 static __rte_always_inline int handle_snat(struct rte_mbuf *m)
 {
 	struct rte_ipv4_hdr *ipv4_hdr;
-	struct rte_tcp_hdr *tcp_hdr;
 	struct dp_flow *df_ptr;
 	struct flow_value *cntrack = NULL;
 	uint32_t src_ip;
@@ -47,11 +46,7 @@ static __rte_always_inline int handle_snat(struct rte_mbuf *m)
 			ipv4_hdr = dp_get_ipv4_hdr(m);
 			ipv4_hdr->src_addr = htonl(dp_get_vm_snat_ip(src_ip, dp_get_vm_vni(m->port)));
 			df_ptr->src.src_addr = ipv4_hdr->src_addr;
-			tcp_hdr =  (struct rte_tcp_hdr *)(ipv4_hdr + 1);
-			ipv4_hdr->hdr_checksum = 0;
-			ipv4_hdr->hdr_checksum = rte_ipv4_cksum(ipv4_hdr);
-			tcp_hdr->cksum = 0;
-			tcp_hdr->cksum = rte_ipv4_udptcp_cksum(ipv4_hdr, tcp_hdr);
+			dp_nat_chg_ip(df_ptr, ipv4_hdr);
 
 			/* Expect the new destination in this conntrack object */
 			cntrack->flow_status = DP_FLOW_STATUS_SRC_NAT;
@@ -68,11 +63,7 @@ static __rte_always_inline int handle_snat(struct rte_mbuf *m)
 		ipv4_hdr = dp_get_ipv4_hdr(m);
 		ipv4_hdr->src_addr = htonl(cntrack->flow_key[DP_FLOW_DIR_REPLY].ip_dst);
 		df_ptr->src.src_addr = ipv4_hdr->src_addr;
-		tcp_hdr =  (struct rte_tcp_hdr *)(ipv4_hdr + 1);
-		ipv4_hdr->hdr_checksum = 0;
-		ipv4_hdr->hdr_checksum = rte_ipv4_cksum(ipv4_hdr);
-		tcp_hdr->cksum = 0;
-		tcp_hdr->cksum = rte_ipv4_udptcp_cksum(ipv4_hdr, tcp_hdr);
+		dp_nat_chg_ip(df_ptr, ipv4_hdr);
 	}
 
 	if (((cntrack->flow_status == DP_FLOW_STATUS_DST_NAT) || (cntrack->flow_status == DP_FLOW_STATUS_DST_LB))
@@ -80,11 +71,7 @@ static __rte_always_inline int handle_snat(struct rte_mbuf *m)
 		ipv4_hdr = dp_get_ipv4_hdr(m);
 		ipv4_hdr->src_addr = htonl(cntrack->flow_key[DP_FLOW_DIR_ORG].ip_dst);
 		df_ptr->src.src_addr = ipv4_hdr->src_addr;
-		tcp_hdr =  (struct rte_tcp_hdr *)(ipv4_hdr + 1);
-		ipv4_hdr->hdr_checksum = 0;
-		ipv4_hdr->hdr_checksum = rte_ipv4_cksum(ipv4_hdr);
-		tcp_hdr->cksum = 0;
-		tcp_hdr->cksum = rte_ipv4_udptcp_cksum(ipv4_hdr, tcp_hdr);
+		dp_nat_chg_ip(df_ptr, ipv4_hdr);
 	}
 	return 1;
 }
